Add more modules for cleanning our codes
