+++ src/index.html
@@ -18,12 +18,10 @@
 
     <main class="container">
         <div class="row" id="contents">
-            <div class="col-2" id="logo"><img
-                    src="https://png.pngtree.com/png-clipart/20200826/original/pngtree-movie-logo-movie-letter-v-png-image_5469427.jpg"
-                    alt="logo"></div>
-            <div class="col-3"><a href="#">spaceships</a></div>
-            <div class="col-3">Planets</div>
-            <div class="col-3">Races</div>
+
+                    
+            <!--Data  from js which counts number of movie items -->
+        
         </div>
